LPS-69436 Avoid major change by deprecating the old taglib. We have agreed with @jbalsas that we will use the new package soy for the new soy taglibs
