+++ modules/apps/collaboration/flags/flags-taglib/package.json
@@ -9,5 +9,5 @@
 		"metal-cli": "^2.1.2"
 	},
 	"name": "flags-taglib",
-	"version": "3.0.0"
+	"version": "2.0.4"
 }
