+++ modules/apps/collaboration/flags/flags-taglib/src/main/resources/META-INF/liferay-flags.tld
@@ -12,7 +12,7 @@
 	<tag>
 		<description>Creates a flag icon that lets users report inappropriate content.</description>
 		<name>flags</name>
-		<tag-class>com.liferay.flags.taglib.servlet.taglib.FlagsTag</tag-class>
+		<tag-class>com.liferay.flags.taglib.servlet.taglib.soy.FlagsTag</tag-class>
 		<body-content>empty</body-content>
 		<attribute>
 			<description>The flagged asset's Java class name.</description>
