+++ modules/apps/collaboration/flags/flags-taglib/src/main/resources/com/liferay/flags/taglib/servlet/taglib/packageinfo
@@ -1,1 +1,1 @@
-version 2.0.0
+version 1.0.0
